+++ CHANGES.md
@@ -1,2 +1,11 @@
 #### [unreleased]
 * initial pass
+* add rows for both update entire site or update all sites with specific repository
+* add ability to use multiple JSON files.
+* update for use with downloadable JSON files from GitHub Updater
+* add javascript to switch display from repos/sites
+* add update feedback messaging
+* create JSON storage directory in `wp-content/uploads/git-remote-updater`
+* sort repositories by name
+* more error checking during update process
+* rebrand to Git Remote Updater